third_party/WebKit/LayoutTests/editing/execCommand/format_block/unrooted-selection-start-crash.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../assert_selection.js"></script>
<script>
test(() => assert_selection(
  '<span contenteditable>^<b><img></b><br>|</span>',
  'justifyCenter',
  [
    '<span contenteditable>',
      '<div style="text-align: center;">',
        '<b>^<img>|</b>',
      '</div>',
     '</span>',
  ]),
  'justifyCenter should not crash');
</script>
